final_prediction_script/requirements.txt
joblib==1.3.2
networkx==3.2
numpy==1.26.4
pandas==2.2.1
pyintergraph==1.3.3
python_igraph==0.11.4
scikit_learn==1.4.1.post1
scipy==1.13.0
torch==2.1.0
torch_geometric==2.5.2
xgboost==2.0.3
